# docs/apispecs.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Kasir App",
        "description": "Kasir App API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/v1"
        }
    ],
    "paths": {
...
        "/user/register": {
            "post": {
                "tags": [
                    "User API"
                ],
                "description": "Register User",
                "summary": "Register User",
                "operationId": "registerUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "number",
                                        "pattern": "^08[0-9]{9,}$"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "confirm_password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/user_response"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/suppliers": {
            "post": {
                "tags": [
                    "Suppliers API"
                ],
                "description": "Create Supplier",
                "summary": "Create Supplier",
                "operationId": "createSupplier",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "number",
                                        "pattern": "^08[0-9]{9,}$"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Supplier Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/suppliers_response"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Suppliers API"
                ],
                "description": "Get Suppliers",
                "summary": "Get Suppliers",
                "operationId": "getSuppliers",
                "responses": {
                    "200": {
                        "description": "Suppliers Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "$ref": "#/components/schemas/suppliers_response"
                                            }
                                            
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/suppliers/{supplier_id}": {
            "put": {
                "tags": [
                    "Suppliers API"
                ],
                "description": "Update Supplier",
                "summary": "Update Supplier",
                "operationId": "updateSupplier",
                "parameters": [
                    {
                        "name": "supplier_id",
                        "in": "path",
                        "description": "Supplier ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "number",
                                        "pattern": "^08[0-9]{9,}$"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Supplier Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/suppliers_response"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Suppliers API"
                ],
                "description": "Delete Supplier",
                "summary": "Delete Supplier",
                "operationId": "deleteSupplier",
                "parameters": [
                    {
                        "name": "supplier_id",
                        "in": "path",
                        "description": "Supplier ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
            "schemas": {
...
                "user_response": {
                    "type": "object",
                    "properties": {
                        "email": {
                            "type": "string"
                        },
                        "phone": {
                            "type": "number"
                        },
                        "created_at": {
                            "type": "string"
                        }
                    }
                },
...
                "suppliers_response": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "number"
                        },
                        "full_name": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "phone": {
                            "type": "number"
                        },
                        "created_at": {
                            "type": "string"
                        }
                    }
                },
...
            },
            "securitySchemes": {
                    "BearerAuth": {
                        "type": "http",
                        "scheme": "bearer",
                        "bearerFormat": "JWT"
                    }
                }
        }
}